X64: Fixups in heap-test.cc to make it compile in X64 mode.
Review URL: http://codereview.chromium.org/113574


git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@1994 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ test/cctest/test-heap.cc
@@ -315,7 +315,7 @@ static bool WeakPointerCleared = false;
 static void TestWeakGlobalHandleCallback(v8::Persistent<v8::Value> handle,
                                          void* id) {
   USE(handle);
-  if (1234 == reinterpret_cast<int>(id)) WeakPointerCleared = true;
+  if (1234 == reinterpret_cast<intptr_t>(id)) WeakPointerCleared = true;
 }
 
 
@@ -385,7 +385,7 @@ TEST(WeakGlobalHandlesMark) {
 static void TestDeleteWeakGlobalHandleCallback(
     v8::Persistent<v8::Value> handle,
     void* id) {
-  if (1234 == reinterpret_cast<int>(id)) WeakPointerCleared = true;
+  if (1234 == reinterpret_cast<intptr_t>(id)) WeakPointerCleared = true;
   handle.Dispose();
 }
